Use Module::Load for optional dependencies and add stats
Replace the call to require with Module::Load's load() function.  This
is a bit less aggressive at trying to satisfy dependencies and won't
re-import if the module is already loaded.

Changed a lot of "TODO: logs" to collect stats instead.  This should
provide visibility without burdening the code with a ton of callbacks.